requirements.txt
streamlit==1.41.1
requests==2.32.3
python-dateutil==2.9.0.post0
